refactor: timeBox MappedSuperClass 제거
