+++ src/main/java/com/debatetimer/dto/customize/request/CustomizeTableCreateRequest.java
@@ -2,7 +2,6 @@ package com.debatetimer.dto.customize.request;
 
 import com.debatetimer.domain.TimeBoxes;
 import com.debatetimer.domain.customize.CustomizeTable;
-import com.debatetimer.domain.customize.CustomizeTimeBox;
 import com.debatetimer.domain.member.Member;
 import jakarta.validation.Valid;
 import java.util.List;
@@ -18,7 +17,7 @@ public record CustomizeTableCreateRequest(
         return info.toTable(member);
     }
 
-    public TimeBoxes<CustomizeTimeBox> toTimeBoxes(CustomizeTable customizeTable) {
+    public TimeBoxes toTimeBoxes(CustomizeTable customizeTable) {
         return IntStream.range(0, table.size())
                 .mapToObj(i -> table.get(i).toTimeBox(customizeTable, i + 1))
                 .collect(Collectors.collectingAndThen(Collectors.toList(), TimeBoxes::new));
