+++ src/main/java/com/debatetimer/exception/errorcode/ClientErrorCode.java
@@ -1,6 +1,5 @@
 package com.debatetimer.exception.errorcode;
 
-import com.debatetimer.domain.DebateTimeBox;
 import com.debatetimer.domain.customize.CustomizeTable;
 import com.debatetimer.domain.customize.CustomizeTimeBox;
 import lombok.Getter;
@@ -31,7 +30,7 @@ public enum ClientErrorCode implements ResponseErrorCode {
     ),
     INVALID_TIME_BOX_SPEAKER_LENGTH(
             HttpStatus.BAD_REQUEST,
-            "발언자 이름은 1자 이상 %d자 이하여야 합니다.".formatted(DebateTimeBox.SPEAKER_MAX_LENGTH)
+            "발언자 이름은 1자 이상 %d자 이하여야 합니다.".formatted(CustomizeTimeBox.SPEAKER_MAX_LENGTH)
     ),
     INVALID_TEAM_NAME_LENGTH(
             HttpStatus.BAD_REQUEST,
