+++ src/main/java/com/debatetimer/repository/customize/CustomizeTimeBoxRepository.java
@@ -22,9 +22,9 @@ public interface CustomizeTimeBoxRepository extends Repository<CustomizeTimeBox,
 
     List<CustomizeTimeBox> findAllByCustomizeTable(CustomizeTable table);
 
-    default TimeBoxes<CustomizeTimeBox> findTableTimeBoxes(CustomizeTable table) {
+    default TimeBoxes findTableTimeBoxes(CustomizeTable table) {
         List<CustomizeTimeBox> timeBoxes = findAllByCustomizeTable(table);
-        return new TimeBoxes<>(timeBoxes);
+        return new TimeBoxes(timeBoxes);
     }
 
     @Query("DELETE FROM CustomizeTimeBox ctb WHERE ctb IN :timeBoxes")
